Add inviteToken logic to the front-end
I'm not proud of this solution, but it will work in a pinch. Let me know
if you have a better idea for this @can-007.

Just using `AUTO_INCREMENT`ing IDs for `url_token` would almost be
enough, except it breaks anonymity - invite codes would then reveal the
enrollment order of users.

We could hash an `AUTO_INCREMENT`ing ID, but handling collisions would
be a bit of a pain.

Maybe this is fine. What do you think?