Seed the 2019 Pilot with challenge+team+achievement data
This commit finalizes the pilot name, competing teams, and available
achievements. As we're starting QA next week, it is important to have
these in place on all of our DBs.

The repeated subqueries aren't pretty, but it reads nicely as a template
string and we don't need to worry about performance on such a small job.