Fix unique key warning in challenge leaderboards
React Fragments have some undefined interactions with lists and keys,
and weren't designed to contain dynamic elements like a leaderboard
list.

This commit reverts our move to fragments back to the nested list
pattern from `leaderboard-card.tsx`.

In the future, these components could be refactored to be clearer and more
efficient. Since the structure of a leaderboard will always be:

1. FetchRow?
2. ...items
3. FetchRow?

Both files should be refactored to render FetchRow outside of the
`items` list.

This also raises the question: why aren't challenge lists reusing the
`leaderboard-card` component?

I've added TODOs for the future; this commit fixes the immediate issue.