RFC: Add validated_at column to clips table
To reduce server load, we want to cache some dynamic properties on clips so we
don't need to query the votes table as often. We're already storing `is_valid`,
but knowing _when_ validation happens is also important.

This RFC includes an example of how existing queries could be simplified. See
the modified query in getClipsStats().