UI[web]: Center Get Involved sectionfor tablet viewport size
The Get Invloved section of the About page wasn't centered once
the viewport width dropped below 768px and flex-direction changed
as a result of media query. The section snapped to the left. The
UI inconsistency isn't noticable much on desktop or mobile device
but its much more pronounced in tablets.
Adding width:100% expands the div to the maximum width effectively
centering the section.