Proposal: enable audio recording for iOS
Is a polyfill really all it takes? Doubt it, but we're gonna find out.

This is untested work, and I'm pushing it as more of an RFC at this
point. The other reason I'm pushing it earlier than usual is that we
can only test this on staging. Details:

This is really hard to test locally. In insecure contexts, including
serving over a non-SSL'd local devserver, `navigator.mediaDevices` is
`undefined`. Going to <localIP>:9000/speak from the devserver itself
doesn't even work. So, testing microphone support locally on a phone is
a non-starter. See
https://developer.mozilla.org/en-US/docs/Web/API/MediaDevices/getUserMedia#Security
for further details.