Add a survey modal to the /datasets page (#2625)
* Add a survey modal to the /datasets page

Brings back the modal that was removed in 8f50819a6823, with updated
language and link.  Also reformats the component to be a bit more
screen reader friendly.

Note for reviewer:

I removed two unneeded SVG attributes from an unrelated icon in this
commit. They were throwing console errors on /datasets, and I didn't
think the fix merited a standalone commit.

* Update copy