Add a note about Docker's memory settings
Auto-rebuilding from `/web` has been broken on my machine for the past
week. I had time to look into it this afternoon. Turns out Docker was
running out of memory and crashing with a cryptic error.

I was already allocating 2GB to Docker, so hopefully we can continue to
reduce the memory footprint of `voice-web`. Until then, I thought this
would be a useful note for folks running the app for the first time.