Update the CSP header
Our audio-recorder-polyfill creates a worker for media encoding, and
creates a blob within the worker. For this to work with our strict CSP,
we need to explicitly allow it.

Explanation:

- Added `'self'` declaration to the `child-src` policy to restrict to
  local worker-created blobs (relevant for recording)
- Added `blob:` scheme source to `connect-src` to allow playback of
  worker-created audio blobs

Reference to the previously erroring line of code, fixed with this
commit (recording bug):

https://github.com/ai/audio-recorder-polyfill/blob/d42c6d0a164f6c02c8e136147c4fd8838c124352/index.js#L53

Also fixes a playback bug, where recorded audio playback was disallowed.