Enable audio recording in iOS Safari (#2680)
* Add an Audio Recorder Polyfill for iOS

Is a polyfill really all it takes? Nope! See follow-on commits for
updates to our audio code to get it running on iOS.

I pushed this early to start testing on staging ASAP. More details below
for the curious:

This is really hard to test locally. In insecure contexts, including
serving over a non-SSL'd local devserver, `navigator.mediaDevices` is
`undefined`. Going to <localIP>:9000/speak from the devserver itself
doesn't even work. So, testing microphone support locally on a phone is
a non-starter. See
https://developer.mozilla.org/en-US/docs/Web/API/MediaDevices/getUserMedia#Security
for further details.

* Make audio-web.js iOS Safari friendly

A few changes:

1) `webkit`-prefix the `AudioContext`
2) Use `(add|remove)EventListener` instead of directly setting
   `recorder.onstart`, etc.
3) Record .wav on iOS, since .ogg is unsupported
4) Use an explicit <source> tag inside of <audio>, since it's required
   for iOS playback

* Misc cleanup

* Add a custom /speak error state for iOS third-party browsers

* Remove unnecessary updateVolume call

* Improve iOS browser sniffing utilities

- Rename isMobileWebkit() to isMobileSafari() for accuracy.
- Improve isMobileSafari() regex to exclude Firefox, Focus, and Mercury – not just Chrome and Opera.
- Remove isFirefoxFocus(): it was only being used to append an unused `.focus` class to the document body, and for a redundant "unsupported platform" check.

* Fix audio-web.ts's clear() method

The move from lastRecordingData and lastRecordingUrl to a single "last"
object in 6cbf5b8772ed2a55 was completely transitioned for audio-ios.ts,
but not for audio-web.ts.

* Update the CSP header

Our audio-recorder-polyfill creates a worker for media encoding, and
creates a blob within the worker. For this to work with our strict CSP,
we need to explicitly allow it.

Explanation:

- Added `'self'` declaration to the `child-src` policy to restrict to
  local worker-created blobs (relevant for recording)
- Added `blob:` scheme source to `connect-src` to allow playback of
  worker-created audio blobs

Reference to the previously erroring line of code, fixed with this
commit (recording bug):

https://github.com/ai/audio-recorder-polyfill/blob/d42c6d0a164f6c02c8e136147c4fd8838c124352/index.js#L53

Also fixes a playback bug, where recorded audio playback was disallowed.

* Choose audio format based directly on browser compatibility