Remove the "last" reference from audio-web.ts
Due to some overaggressive revoking of URLs, we were seeing bugs in
replaying and re-recording audio on the /speak page (eg. #2716, and
perhaps related to #2553). `this.last` was a huge vector for bugs, and
only used for `URL.revokeObjectURL` calls, so this commit moves URL
cleanup to `speak.tsx` and defers it until a Skip or Upload action
occurs.

This commit contains various other cleanups and small fixes, such as
making sure the most recent version of state.clips is referenced in our
setState calls. We paired on this together, so I won't go into any
further detail :)