Proof of concept: associate demographic metadata with clips
We create clip metadata on the fly for each dataset release, based on the speaker’s demographic info available at that time. This creates a problem for long-term users who need to change their profile’s age, sex, or accents; all previous clips will be updated with new, inaccurate metadata.

This commit is a first step toward addressing this problem. It creates a new `demographics` table to hold the demographic history of our users. Clips can be locked to a specific entry in this table, so we don’t risk changing metadata in the future.

As discussed offline, there are a few inelegant hacks going on here. To avoid locking up our DB with a huge backfill on `clips` and `user_clients`, I’ve created two one-to-one tables:

```
CREATE TABLE user_client_demographics…
CREATE TABLE clip_demographics…
```

These respectively fake an additional `demographics` column on `user_clients` and on `clips`. I don’t like it, and if you have any ideas about how we can backfill our tables without taking the site down I’d be happy to do that instead (or try something else entirely).

Accents are a bit of a pickle, so I’m keeping them out of this commit. Profiles can have multiple languages, so there may be a one-to-many relationship between users and accents. It will probably be easiest to keep track of historical accent data in a separate table, and I thought there was already enough going on here for a proof of concept. If it gets the go-ahead, I’ll start moving on accents.