Don't rely on client's locale when saving clips
We've been seeing reports of clips of one language showing up on the
/listen page of another language. Our client does its best to be
responsive and quick when a user changes their language… as it should.
But when someone changes their language mid-way through recording a
batch of clips, it shouldn't affect the language those clips are saved
under.

This is a small and simple commit to move locale-handling to the server.
It adds a db call with `db.findSentence`, but removes three calls by
avoiding `db.getLocaleId`… so I think this might be a performance win in
the end.