Associate demographic metadata with clips (#2868)
* Proof of concept: associate demographic metadata with clips

We create clip metadata on the fly for each dataset release, based on the speaker’s demographic info available at that time. This creates a problem for long-term users who need to change their profile’s age, sex, or accents; all previous clips will be updated with new, inaccurate metadata.

This commit is a first step toward addressing this problem. It creates a new `demographics` table to hold the demographic history of our users. Clips can be locked to a specific entry in this table, so we don’t risk changing metadata in the future.

As discussed offline, there are a few inelegant hacks going on here. To avoid locking up our DB with a huge backfill on `clips` and `user_clients`, I’ve created two one-to-one tables:

```
CREATE TABLE user_client_demographics…
CREATE TABLE clip_demographics…
```

These respectively fake an additional `demographics` column on `user_clients` and on `clips`. I don’t like it, and if you have any ideas about how we can backfill our tables without taking the site down I’d be happy to do that instead (or try something else entirely).

Accents are a bit of a pickle, so I’m keeping them out of this commit. Profiles can have multiple languages, so there may be a one-to-many relationship between users and accents. It will probably be easiest to keep track of historical accent data in a separate table, and I thought there was already enough going on here for a proof of concept. If it gets the go-ahead, I’ll start moving on accents.

* Respond to feedback from #2764

- Fix some errors and omissions
- s/gender/sex
- Create separate tables for age and sex values

* Consolidate demographics migration

* Respond to feedback from #2764

- Remove the hack-y `user_client_demograpics` table, find the most
  recent entry dynamically with subqueries
- Fix some bugs in the schema definition - not sure how those found
  their way into the initial commit

* INSERT to clip_demographics for new clip uploads

* s/sex/gender

* Improve age/gender matching from prior deprecated fields

Added TRIM and LOWER, just in case. I also switched:

```
        WHERE user_clients.deprecated_age IS NOT NULL
           OR user_clients.deprecated_gender IS NOT NULL
```

with:

```
        WHERE ages.age IS NOT NULL
           OR genders.gender IS NOT NULL
```

to ensure that accounts with invalid and unsalvageable demographic entries don't make it into this table.

---

(I also fixed all invalid ages / genders in our DB, but just in case something breaks before this hits prod)

* Handle undefined or invalid values in updateDemographics

* Add logging on a likely unused method