Prevent blank clips sent to the server from propagating
Hopefully over time we send less and less blank clips from the client.
But until every client and device is perfectly reliable, we should also
double-check that clips are valid on the server. For now, we're using
file size as a proxy for that.

After this commit, this is the flow of an audio clip from the client, to
the server, to AWS, to the DB:

- Clip is recorded on the client
- Clip is submitted to the server
- HTTP stream of the clip is streamed to ffmpeg, which is then piped to
  an AWS streaming upload
- Once the upload and transcoding are complete, if the audio clip is
  invalid or resulted in an mp3 size < 300 bytes, we bail and delete the
  file from AWS
- Otherwise, we consider it valid and save it to the DB

Specific changes in this commit:

- Update audio-recording-polyfill to hopefully reduce incidence of blank
  clips from the client
- Add ^that whole error-checking flow^ on the server
- Add a few transcoding options to ffmpeg to ensure consistent #
  channels and sample rate.