Fix a race condition when updating demographics data
This is caused by not awaiting the updateDemographics() promise. It
it therefore scheduled for running, but may be executed *after* the
UserClient.findAccount() call is awaited. Therefore, there is race
between updating the demographics and querying for it later, resulting
in a non-deterministic retrieval and the gender/age data sometimes
disapearing from the form. This is confusing for the user who thinks
the data was not saved.

I've seen it happening on the prod website.